Menu Play button bug fixed
playbutton wouldnt work after switching scenes. this was due to the difficultymanager script that used "DontDestroyOnLoad(gameObject)" was also on the same object as the menuhandler script. This resulted in the difficultymanger script being saved but destroying the menuhandle script and anything else in the same object. This was fixed by sepereating the difficultymanger script into its own object.